Refactored OpeningScreen class to its separate file.
